+++ lib/shrine/storage/file_system.rb
@@ -67,9 +67,9 @@ class Shrine
     # Note that this approach has a couple of downsides. For example, you can
     # only use it for cache, since Heroku wipes this directory between app
     # restarts. This also means that deploying the app can cancel someone's
-    # uploading if you're using backgrounding. Also, you cannot display a link
-    # to the cached file before it is promoted to store, since it won't be in
-    # the "public" directory.
+    # uploading if you're using backgrounding. Also, by default you cannot
+    # generate URLs to files in the "tmp" directory, but you can with the
+    # download_endpoint plugin.
     class FileSystem
       attr_reader :directory, :subdirectory, :host, :permissions